Lee jy

+++ the_joyful_delivery/src/main/java/DTO/Delivery.java
@@ -21,7 +21,7 @@ import lombok.NoArgsConstructor;
 public class Delivery {
 	
 	@Pk
-	private int idx;
+	private Long idx;
 	private int userIdx;
 	private int driverIdx;
 	private int srIdx;
@@ -38,6 +38,7 @@ public class Delivery {
 	private String regStatus;
 	@Column(exist = false)
 	private String name;
+
 	@Column(exist = false)
 	private SendingReceiving sendingReceiving; // 조인해서 가져올 SendingReceiving 객체
 	@Column(exist = false)
